+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/product/list/ProductsViewModel.kt
@@ -1,0 +1,69 @@
+package com.app.arcabyolimpo.presentation.screens.product.list
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.app.arcabyolimpo.domain.common.Result
+import com.app.arcabyolimpo.domain.model.product.Product
+import com.app.arcabyolimpo.domain.usecase.product.GetProductsListUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+data class ProductsUiState(
+    val isLoading: Boolean = false,
+    val error: String? = null,
+    val products: List<Product> = emptyList(),
+    val searchQuery: String = ""
+)
+
+@HiltViewModel
+class ProductsViewModel @Inject constructor(
+    private val getProductsListUseCase: GetProductsListUseCase
+) : ViewModel() {
+
+    private val _uiState = MutableStateFlow(ProductsUiState())
+    val uiState: StateFlow<ProductsUiState> = _uiState
+
+    init {
+        loadProducts()
+    }
+
+    private fun loadProducts() {
+        viewModelScope.launch {
+            getProductsListUseCase().collect { result ->
+                when (result) {
+                    is Result.Loading -> {
+                        _uiState.update { it.copy(isLoading = true, error = null) }
+                    }
+
+                    is Result.Success -> {
+                        _uiState.update {
+                            it.copy(
+                                isLoading = false,
+                                error = null,
+                                products = result.data
+                            )
+                        }
+                    }
+
+                    is Result.Error -> {
+                        _uiState.update {
+                            it.copy(
+                                isLoading = false,
+                                error = "Error al cargar productos"
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+
+    fun onSearchQueryChange(newQuery: String) {
+        _uiState.update { it.copy(searchQuery = newQuery) }
+    }
+}