+++ CHANGELOG.md
@@ -1,14 +1,20 @@
-## [0.1.0] - 15/11/2025
+## [0.1.0] - 16/11/2025
 
 ### Added
-- Added US to Delete Product Batches
+ - Added products search
+ - Add DeleteWorkshop Logic, Data and Domain
+ - Add way to get users for workshops
+ - Added functionality search beneficiaries
+ - Filter and order beneficiaries functionality
+ - Fixed merge conflicts
+ - Added US to Delete Product Batches
 
 ### Fixed
 - Refreshing on Product Batches List
 
 ## [0.1.0] - 15/11/2025
 ### Fixed
-- Fixed error from standard input attibutes into register supply batch screen
+
 
 ## [0.1.0] - 15/11/2025
 
@@ -18,6 +24,7 @@
 
 ### Fixed
 - Fixed StandardInput missing parameter
+- Fixed error from standard input attibutes into register supply batch screen
 
 ## [0.1.0] - 14/11/2025